+++ packages/components/src/utils/hooks.ts
@@ -34,3 +34,28 @@ export const useKeyPress = (targetKey: string) => {
 
     return keyPressed;
 };
+
+export const useOnClickOutside = (
+    ref: React.MutableRefObject<HTMLDivElement | null>,
+    callback: (event: MouseEvent | TouchEvent) => void
+) => {
+    useEffect(() => {
+        if (!ref) return;
+        const listener = (event: MouseEvent | TouchEvent) => {
+            // Do nothing if clicking ref's element or descendent elements
+            if (!ref.current || ref.current.contains(event.target as Node)) {
+                return;
+            }
+
+            callback(event);
+        };
+
+        document.addEventListener('mousedown', listener);
+        document.addEventListener('touchstart', listener);
+
+        return () => {
+            document.removeEventListener('mousedown', listener);
+            document.removeEventListener('touchstart', listener);
+        };
+    }, [ref, callback]);
+};