+++ packages/suite/src/actions/suite/constants/suiteConstants.ts
@@ -15,6 +15,7 @@ export const FORGET_DEVICE = '@suite/forget-device';
 export const REMEMBER_DEVICE = '@suite/remember-device';
 export const SET_LANGUAGE = '@suite/set-language';
 export const SET_DEBUG_MODE = '@suite/set-debug-mode';
+export const SET_FLAG = '@suite/set-flag';
 export const ONLINE_STATUS = '@suite/online-status';
 export const APP_CHANGED = '@suite/app-changed';
 export const ADD_BUTTON_REQUEST = '@suite/add-button-request';